Set debugger build to use -O2 - much faster.
Fixed issue with missing ruby in r594, a lot of the applications and libraries needed were being cmake'd - these
have now been moved into their own directory tree which is copied to the disk in create_images.sh.

I may have borked the bashrc, it doesn't look as nice as before and has no $PATH. This will be changed back in the next revision.