Got the event system working - for kernel mode handlers, anyway.
Tested with a proof-of-concept class TimeoutGuard. This class is based
on RAII, just like LockGuard. Construct one with a timeout in its
constructor, and if it isn't destroyed before that timeout finishes, it
will jump back to the statement after the constructor.

Therefore, create one thus:

{
	TimeoutGuard guard(2);
	if (!guard.timedOut())
	{
		// Do possibly blocking statements.
	}
}
// Will get here in 2 seconds maximum.

Proper documentation in markdown format for the event system is
forthcoming.