winman: start /applications/tui whenever the user hits <mod>-ENTER.
So, this breaks in awesome ways with the way the TUI works. You can end up
with several TUI instances hooked up to the same UserConsole, and neither of
them function very well in this case. To make this work, we'll need to
do a lot of work in the TUI syscalls - this will be happening shortly.