winman: provide a 'sync' message, which is used by Widget::getFramebuffer.
This means that before any frame is rendered, the client ensures the window
manager is prepared for this rendering to take place. The window manager
provides a GraphicsProvider in the response which may not be the same as
in the current Widget framebuffer. The Widget layer handles all this. Sadly,
it turns out that this doesn't resolve the rendering issues with the TUI
during modeChanged to a smaller framebuffer. Still debugging...