SDL: initial cut of integration with the window manager.
This is horribly broken right now. 'fire' for example complains that it
didn't get a display mode big enough for its purposes. I'd like it to clip
to the tile it's running within, if it cannot resize itself.

That is, if SDL_RESIZABLE is set on the flags, we can probably just use
the raw window framebuffer. If not, we should try and clip.

Needs some more thinking...