build, x86_64: output a 32-bit ELF binary on x86_64, which allows booting with unpatched GRUB.
This has required some changes to KernelElf to handle the fact that the kernel ELF is a 32-bit ELF object, and therefore addresses have been truncated.
Also, this has been done with a 32-bit assembly stub that enters long mode and then jumps to the normal 64-bit assembly entry point.