posix: don't use userspace stack too early when creating pthreads
So, this actually fixes a really horrible race condition. Processor::jumpKernel has a 'sti' as its final step before jumping to the new code to enable interrupts. This is fine, except that when we use the userspace stack as our stack, that 'sti' can come with an interrupt & task switch. The task switch will enter the new address space and then attempt to pop the old %ebp (design decision) from a stack that is not mapped. This then triggers a #PF in ring0, which then triggers a #DF as there's no stack on which to actually handle the #PF.

Using the correct kernel stack for the kernel side of the thread before jumping to userspace with the correct stack solves this.