posix, boot, core: tweak optimisations via #pragma for mem* functions
I'm OK with the compiler turning patterns that look like memset/memcpy into calls to memset/memcpy, but it can't do so in these particular files. So, rather than force the compiler to stop doing that we'll fix this up with a pragma.

This should be a big step towards being able to run the ARM port at -O3.