process: fix race condition in Semaphore
The check for pending threads to wake up in release() was done without the modification lock. On SMP systems, that allowed (in some cases) for the queue to appear empty, even though another CPU was busy populating it. The end result is the other CPU's thread goes into a permanent sleep, which is bad.

This fixes that case.