debugger: further extenedd LocksCommand, including rewriting the debugger CLI command
The command now shows a view of every lock acquired by each CPU, visually nested to indicate the depth of acquire()s.

One more thing that'd be great is to reintroduce the backtraces for each log entry and have the ability to show them in the UI, which can help identify the code path to a particular failure.