kernel: revamp logging to reduce the scope of the Log lock
In some cases, the actual act of reading e.g. a string to print can cause an exception (a trap to map in the page, for example), which could lead to the core becoming alive and having interrupts enabled even while the spinlock is active. By performing the actual creation of the log string separately, the actual part that needs the lock (storing the log entry) can be done safely. Any failures/traps will happen during the entry creation, without the Log lock being taken.

This came up thanks to the recent addition of interrupt checks to the lock checker. Some POSIX syscalls log their inputs, and that can cause a trap which re-enables interrupts to read from disk.